Fix main.rb
