+++ Books/JSON_files/books.json
@@ -1,1 +1,1 @@
-[{"id":517,"publisher":"DSF","cover_state":16,"publish_date":"sdfsd","archived":"good","label":"2022-01-01"},{"id":"jonh","publisher":"good","cover_state":"2022-09-09","publish_date":589,"archived":"JJJ","label":110},{"id":920,"publisher":"Nama","cover_state":"good","publish_date":"2022-09-09","archived":854,"label":"JJJ"},{"id":"2022-10-10","publisher":421,"cover_state":"S","publish_date":609,"archived":"sdf\\","label":"good"}]
+[{"id":"good","publisher":"2022-01-01","cover_state":517,"publish_date":"DSF","archived":16,"label":"sdfsd"},{"id":"JJJ","publisher":110,"cover_state":"jonh","publish_date":"good","archived":"2022-09-09","label":589},{"id":854,"publisher":"JJJ","cover_state":920,"publish_date":"Nama","archived":"good","label":"2022-09-09"},{"id":"sdf\\","publisher":"good","cover_state":"2022-10-10","publish_date":421,"archived":"S","label":609},{"id":"DFSFS","publisher":624,"cover_state":"testfinal","publish_date":"good","archived":"2022-10-10","label":false}]
